update: website ban quan ao

+++ src/main/java/com/example/myApp/controller/ProductController.java
@@ -6,6 +6,7 @@ import com.example.myApp.dto.ProductResponse;
 import com.example.myApp.enity.Products;
 import com.example.myApp.repository.CategoryRepository;
 import com.example.myApp.repository.ProductRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
@@ -61,10 +62,14 @@ public class ProductController {
     }
 
     @PostMapping(value = "/add", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<?> addProducts(@ModelAttribute ProductDTO productDTO,
-                                         @RequestParam("image") MultipartFile image){
+                                         @RequestParam("image") MultipartFile image,
+                                         HttpServletRequest request){
         try {
+            Integer userId = (Integer) request.getAttribute("userId");
+            if (userId == null || userId != 2) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Bạn không có quyền");
+            }
             Products savedProduct = productService.addProducts(productDTO, image);
             return ResponseEntity.ok(savedProduct);
         } catch (RuntimeException e) {
@@ -75,7 +80,12 @@ public class ProductController {
     @PutMapping("/update/{id}")
     public ResponseEntity<String> updateProducts(
             @PathVariable int id,
-            @RequestBody ProductDTO productDTO){
+            @RequestBody ProductDTO productDTO,
+            HttpServletRequest request){
+        Integer userId = (Integer) request.getAttribute("userId");
+        if (userId == null || userId != 2) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Bạn không có quyền");
+        }
         Products products = productRepository.findById(id).orElseThrow(
                 () -> new RuntimeException("Not found product with id = " + id));
         products.updateFromDTO(productDTO, categoryRepository);
@@ -84,7 +94,12 @@ public class ProductController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<String> deleteProducts(@PathVariable int id){
+    public ResponseEntity<String> deleteProducts(@PathVariable int id,
+                                                 HttpServletRequest request){
+        Integer userId = (Integer) request.getAttribute("userId");
+        if (userId == null || userId != 2) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Bạn không có quyền");
+        }
         productService.deleteProduct(id);
         return ResponseEntity.ok("Delete success");
     }
